+++ posthog/src/main/java/com/posthog/PostHog.kt
@@ -678,16 +678,16 @@ public class PostHog private constructor(
         val except = listOf(VERSION, BUILD)
         getPreferences().clear(except = except)
         featureFlags?.clear()
-        queue?.clear()
-        replayQueue?.clear()
         featureFlagsCalled.clear()
-
         synchronized(identifiedLock) {
             isIdentifiedLoaded = false
         }
 
         endSession()
         startSession()
+
+        // reload flags as anon user
+        reloadFeatureFlags()
     }
 
     private fun isEnabled(): Boolean {